% Generated by roxygen2 (4.0.2): do not edit by hand
\name{component_association_test}
\alias{component_association_test}
\title{Testing the association between independent component coefficients and known covariates.}
\usage{
component_association_test(input.A, info.input, covar.names)
}
\arguments{
\item{A}{The mixing matrix estimated through ICA   \code{A}}

\item{info.df}{A dataframe that holds the measured covariates for each sample \code{info.df}}
}
\value{
output A matrix holding the p-values for each indepenedent component and covariate pair.
}
\description{
The function takes in the A matrix from the ica.result object
with a dataframe of measured covariates to test the association
between them.
}
\keyword{keywords}

