% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_h5_file}
\alias{create_h5_file}
\title{Creating eQTL input hdf5 file templates.}
\usage{
create_h5_file(file_name)
}
\arguments{
\item{file_name}{The name of the file to be created}
}
\value{
HDF5 file with groups phenotypes, genotypes, and covars with subgroups col_info and row_info will be created in the current directory.
}
\description{
Generates an hdf5 file with the basic groups for eQTL analysis.
}
\keyword{HDF5}
\keyword{eQTL,}

