% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fastICA_gene_expr}
\alias{fastICA_gene_expr}
\title{ICA analysis based on the package fastICA}
\usage{
fastICA_gene_expr(X, n.comp, fun = "logcosh", alpha = 1,
  scale.pheno = FALSE, maxit = 200, tol = 1e-04, verbose = TRUE,
  w.init = NULL)
}
\arguments{
\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx. \code{scale.pheno}}

\item{phenotype.mx}{Phenotype matrix with diemnsions g x N   \code{phenotype.mx}}

\item{k.est}{Number of components to be estimated or method to estimate it. \code{k.est}}
}
\value{
List with the following entries.
}
\description{
Minor tweak to fastICA to make it faster
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

