% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene_expr_pca}
\alias{gene_expr_pca}
\title{Custom PCA function for analyzing gene expression data.}
\usage{
gene_expr_pca(phenotype.mx = NULL, info.df = NULL, check.covars = NULL,
  scale.pheno = FALSE, cor.threshold = 0.05, ...)
}
\arguments{
\item{phenotype.mx}{Phenotype matrix with diemnsions g x N}

\item{info.df}{Dataframe that holds sample covariates (ex. population, gender, age, etc...)}

\item{check.covars}{Column names of info.df which hold the covariates
that should be used for association testing with IC coefficients.}

\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx. Default is set to FALSE.}

\item{cor.threshold}{Threshold for significant correlation calling. Default is set to 0.05.}
}
\value{
List with the following entries.
}
\description{
Performing PCA on a dataset and create a list object with results.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

