% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genotypes_pca}
\alias{genotypes_pca}
\title{Custom PCA function for analyzing genotype SNP data.}
\usage{
genotypes_pca(genotype.mx = NULL, info.df = NULL, check.covars = NULL,
  cor.threshold = 0.05)
}
\arguments{
\item{genotype.mx}{Genotype matrix with diemnsions g x N}

\item{info.df}{Dataframe that holds sample covariates (ex. population, gender, age, etc...)}

\item{check.covars}{Column names of info.df which hold the covariates
that should be used for association testing with IC coefficients.}

\item{cor.threshold}{Threshold for significant correlation calling. Default is set to 0.05.}
}
\value{
List with the following entries.
}
\description{
Performing PCA on a dataset and create a list object with results.
}
\keyword{keywords}

