% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggplot_add_theme}
\alias{ggplot_add_theme}
\title{Adding customized black and white theme to ggplot objects and controls the font size.}
\usage{
ggplot_add_theme(inputplot)
}
\arguments{
\item{inputplot}{A ggplot object.  \code{s}}
}
\value{
A ggplot object with a custom theme added
}
\description{
The function takes a ggplot object as an input and modifies the theme of it
}

