% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h5_add_data}
\alias{h5_add_data}
\title{Adding phenotypes to a given hdf5 file}
\usage{
h5_add_data(file_name = NULL, input_matrix = NULL, feature_ids = NULL,
  sample_ids = NULL, data_type = c("phenotypes", "genotypes", "covars"))
}
\arguments{
\item{file_name}{Name of the HDF5 file to add the information to}

\item{feature_ids}{Unique phenotype ids}

\item{sample_ids}{Unique sample ids}

\item{phenotypes}{Phenotype matrix to save}
}
\value{
Saves the given objects into the hdf5 generated by the create_eqtl_input_h5() function.
}
\description{
Use in-memory objects to create phenotypes datasets in a given hdf5 file
}
\keyword{HDF5}
\keyword{eQTL,}

