% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h5_add_snp_info}
\alias{h5_add_snp_info}
\title{Adding SNP information to an HDF5 file}
\usage{
h5_add_snp_info(file_name = NULL, snp_id = NULL, snp_chr = NULL,
  snp_pos = NULL)
}
\arguments{
\item{file_name}{Name of the HDF5 file to add the information to}

\item{snp_id}{Vector of genotype ids that match the feature ids used in saving the genotype matrix.}

\item{snp_chr}{Vector of genotype chromosomes in the same order as snp_id entries.}

\item{snp_pos}{Vector of genotype positions in the same order as snp_id entries.}

\item{snp_info}{Dataframe with SNP information for all SNPs present in genotypes/matrix}
}
\value{
Saves the given objects into the hdf5 generated by the create_eqtl_input_h5() function.
}
\description{
Use in-memory objects to add SNP information to a given HDF5 file
}
\keyword{HDF5}
\keyword{eQTL,}

