% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h5_ica}
\alias{h5_ica}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
h5_ica(h5_file = NULL, input_pheno = NULL, k.est = NULL,
  scale.pheno = FALSE, var.cutoff = 99)
}
\arguments{
\item{h5_file}{Name of HDF5 file that has the phenotype values saved.}

\item{input_pheno}{Phenotype matrix with diemnsions N x g}

\item{k.est}{Number of components to be estimated or method to estimate it.}

\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

