% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ica_covar_check}
\alias{ica_covar_check}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
ica_covar_check(ica_list = NULL, h5_file = NULL, covars = NULL)
}
\arguments{
\item{h5_file}{Name of HDF5 file that has the phenotype values saved.}

\item{input_pheno}{Phenotype matrix with diemnsions N x g}

\item{k.est}{Number of components to be estimated or method to estimate it.}

\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

