% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ica_R_par}
\alias{ica_R_par}
\title{ICA parameter estimation function}
\usage{
ica_R_par(X, n.comp, tol, fun, alpha, maxit, verbose, w.init)
}
\arguments{
\item{phenotype.mx}{Phenotype matrix with diemnsions g x N   \code{phenotype.mx}}

\item{k.est}{Number of components to be estimated or method to estimate it. \code{k.est}}

\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx. \code{scale.pheno}}
}
\value{
List with the following entries.
}
\description{
Adopted from fastICA package
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

