% Generated by roxygen2 (4.0.2): do not edit by hand
\name{k_est_singlerun}
\alias{k_est_singlerun}
\title{Estimating the number of components (k)}
\usage{
k_est_singlerun(phenotype.mx = NULL, max.iter = 10, run.index = 1)
}
\arguments{
\item{phenotype.mx}{The matrix that holds the gene expression profile   \code{phenotype.mx}}

\item{max.iter}{Number of maximum iterations, default is set to 10 \code{max.iter}}

\item{run}{index \code{run.index}}
}
\value{
output A matrix holding the p-values for each indepenedent component and covariate pair.
}
\description{
Estimating the number of components to be estimated
through the replication ratio between two independent runs of ICA
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

