% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_data}
\alias{load_data}
\title{Loading data from a given hdf5 file}
\usage{
load_data(input_file = NULL, data_type = c("phenotypes", "genotypes",
  "covars"))
}
\arguments{
\item{data_type}{Data type to create (phenotypes, genotypes, covars)}

\item{file_name}{Name of the HDF5 file to read from}
}
\value{
Data matrix with row and columns labels.
}
\description{
Load data from a given hdf5 file to an in-memory object
}
\keyword{HDF5}
\keyword{eQTL,}

