% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pc_report}
\alias{pc_report}
\title{Automated PCA report generator}
\usage{
pc_report(pca.result = NULL, plot.pc = NULL, prefix = NULL,
  output.path = NULL, file.ext = "png")
}
\arguments{
\item{pca.result}{PCA result list created by the function prcomp().   \code{pca.result}}

\item{plot.pc}{Number of principal components to plot. Default is set to plot every PC. \code{plot.pc}}

\item{prefix}{Output filename prefix. The output file will be named
"prefix_ICA_summary.html". \code{prefix}}

\item{output.path}{Directory path for generating the output HTML file.
default is set to current working directory. \code{output.path}}

\item{file.ext}{Extension of plots saved in the process of compiling the HTML report.
Default is set to png format. \code{file.ext}}
}
\value{
output HTML report that contains component wise visualization and summary plots.
}
\description{
Generating a HTML report from a PCA object created with the function prcomp().
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

