% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peak_detection}
\alias{peak_detection}
\title{Labeling peaks for each independent component}
\usage{
peak_detection(ica.input.s)
}
\arguments{
\item{s}{A single column of the S matrix with dimensions 1 x g  \code{s}}
}
\value{
list with entries "peaks" and "N". "peaks" has the position of peak genes
in the decreasing order of magnitude. "N" has the total number of peak genes.
}
\description{
The function takes in the A matrix from the ica.result object
with a dataframe of measured covariates to test the association
between them.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

