% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_coeff_w_legend}
\alias{plot_coeff_w_legend}
\title{Plotting the coefficients for a single IC}
\usage{
plot_coeff_w_legend(info_input, k.col, plot.title)
}
\arguments{
\item{plot.title}{Character title of the plot\code{plot.title}}

\item{s}{A single component  \code{s}}
}
\value{
ggplot object
}
\description{
ggplot function for plotting coefficients for a single IC
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

