% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_component}
\alias{plot_component}
\title{Plotting the gene weights of a single component}
\usage{
plot_component(s, plot.title, peaks = FALSE, peakresult = NULL,
  gene.names = NULL)
}
\arguments{
\item{s}{A single component  \code{s}}

\item{plot.title}{Character title of the plot\code{plot.title}}
}
\value{
ggplot object
}
\description{
ggplot function for plotting a single component
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

