% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_component_chr}
\alias{plot_component_chr}
\title{Plotting function that indicates gene position on chromosome}
\usage{
plot_component_chr(s, geneinfo.df, x.axis, plot.title, peaks = TRUE,
  peakresult = NULL)
}
\arguments{
\item{s}{A single component}

\item{geneinfo.df}{Gene position information dataframe}

\item{x.axis}{Positions of the chromosome borders for coloring.}

\item{plot.title}{Character title of the plot}

\item{peaks}{TRUE shows peaks in red, FALSE leaves everything black}

\item{peakresult}{Pre-calculated peak positions, if not supplied peaks will be
       defined as values lying out of 2 standard deviations.}
}
\description{
The function generates a plot with gene weights for individual ICs according
to their position on each chromosome.
}

