% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_component_grid}
\alias{plot_component_grid}
\title{Simplified Plotting function that indicates gene position on chromosome}
\usage{
plot_component_grid(s, geneinfo_df, plot.title)
}
\arguments{
\item{s}{A single component}

\item{geneinfo_df}{Gene position information dataframe with columns "chr"}

\item{plot.title}{Character title of the plot}
}
\description{
The function generates a plot with gene weights for individual ICs according
to their position on each chromosome.
}

