% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_component_hist}
\alias{plot_component_hist}
\title{Plotting a histogram of gene weights of a single component}
\usage{
plot_component_hist(s, plot.title)
}
\arguments{
\item{s}{A single component  \code{s}}

\item{plot.title}{Character title of the plot\code{plot.title}}
}
\value{
ggplot object
}
\description{
ggplot function for plotting a histogram for a single component
}
\keyword{keywords}

