% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_ica}
\alias{run_ica}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
run_ica(pheno_mx = NULL, k_est = NULL, var_cutoff = 99, n_runs = 1,
  n_cores = NULL, scale_pheno = FALSE, h_clust_cutoff = 0.3,
  max_iter = 10, cor_threshold = 0.05, similarity_measure = "peaks")
}
\arguments{
\item{pheno_mx}{Phenotype matrix with diemnsions g x N}

\item{k_est}{Number of components to be estimated or method to estimate it.}

\item{n_runs}{Number of runs for estimating k. Default value is set to 5.}

\item{n_cores}{Number of cores to be used for estimating k. Default is set to 1.}

\item{scale_pheno}{Logical value specifying the scaling of row of the pheno_mx.}

\item{h_clust_cutoff}{is the cutoff value used in hierarchical clustering. Default is set to 0.3.}

\item{max_iter}{Maximum iterations for estimating k for each run. Default value is set to 10.}

\item{cor_threshold}{Threshold for significant correlation calling. Default is set to 0.05.}

\item{similarity_measure}{How to measure the similarity between ICs.
If set to "peaks" only gene weights that are greater than 1 sd are used to calculate similarity.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

