% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tellme_aboutit}
\alias{tellme_aboutit}
\title{Generating ICA or PCA summary reports.}
\usage{
tellme_aboutit(what = NULL, how = "ica", n.comps = NULL, prefix = NULL,
  output.path = NULL, file.ext = "png")
}
\arguments{
\item{what}{ICA or PCA result list created by either \code{gene_expr_ica()} or \code{gene_expr_pca()}.}

\item{how}{Specifying whether the object is a "ica" object or "pca" object. Default is set to "ica".}

\item{n.comps}{Number of principal components to plot. Default is set to plot every PC. Only used for PCA plotting not for ICA plotting.}

\item{prefix}{Output filename prefix. The output file will be named
"prefix_ICA_summary.html".}

\item{output.path}{Directory path for generating the output HTML file.
default is set to current working directory.}

\item{file.ext}{File extension to be used for saved plots. Default is set to png for html reports.
Note that if you use pdf plots for html files they will not show up in your report.}
}
\value{
output HTML report.
}
\description{
Generating a HTML report from a ICA or PCA list object.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

