% Generated by roxygen2 (4.0.2): do not edit by hand
\name{component_association_test}
\alias{component_association_test}
\title{Testing the association between independent component coefficients and known covariates.}
\usage{
component_association_test(input.A, info.input, covar.names)
}
\arguments{
\item{input.A}{The mixing matrix estimated through ICA}

\item{info.input}{A dataframe that holds the measured covariates for each sample}

\item{covar.names}{names of the covariates that should be used for association testing}
}
\value{
output A matrix holding the p-values for each indepenedent component and covariate pair.
}
\description{
The function takes in the A matrix from the ica.result object
with a dataframe of measured covariates to test the association
between them.
}
\keyword{keywords}

