% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene_expr_ica}
\alias{gene_expr_ica}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
gene_expr_ica(phenotype.mx = NULL, info.df = NULL, check.covars = NULL,
  k.est = NULL, scale.pheno = FALSE, h.clust.cutoff = 0.3, n.runs = 5,
  max.iter = 10, n.cores = NULL, cor.threshold = 0.05,
  similarity.measure = "peaks")
}
\arguments{
\item{phenotype.mx}{Phenotype matrix with diemnsions g x N}

\item{info.df}{Dataframe that holds sample covariates (ex. population, gender, age, etc...)}

\item{check.covars}{Column names of info.df which hold the covariates
that should be used for association testing with IC coefficients.}

\item{k.est}{Number of components to be estimated or method to estimate it.}

\item{scale.pheno}{Logical value specifying the scaling of row of the phenotype.mx.}

\item{h.clust.cutoff}{is the cutoff value used in hierarchical clustering. Default is set to 0.3.}

\item{n.runs}{Number of runs for estimating k. Default value is set to 5.}

\item{max.iter}{Maximum iterations for estimating k for each run. Default value is set to 10.}

\item{n.cores}{Number of cores to be used for estimating k. Default is set to 1.}

\item{cor.threshold}{Threshold for significant correlation calling. Default is set to 0.05.}

\item{similarity.measure}{How to measure the similarity between ICs.
If set to "peaks" only gene weights that are greater than 1 sd are used to calculate similarity.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

