% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h5_add_snp_info}
\alias{h5_add_snp_info}
\title{Adding SNP information to an HDF5 file}
\usage{
h5_add_snp_info(file_name = NULL, snp_info = NULL, id_col = "id",
  chr_col = "chromosome", pos_col = "position")
}
\arguments{
\item{file_name}{Name of the HDF5 file to add the information to}

\item{snp_info}{Dataframe with SNP information for all SNPs present in genotypes/matrix}

\item{id_col}{Column name of snp_info that contains id information}

\item{chr_col}{Column name of snp_info that contains chromosome information}

\item{pos_col}{Column name of snp_info that contains position information}
}
\value{
Saves the given objects into the hdf5 generated by the create_eqtl_input_h5() function.
}
\description{
Use in-memory objects to add SNP information to a given HDF5 file
}
\keyword{HDF5}
\keyword{eQTL,}

