% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ic_report}
\alias{ic_report}
\title{Automated ICA report generator}
\usage{
ic_report(ica.result, prefix, output.path = NULL, file.ext = "png")
}
\arguments{
\item{ica.result}{ICA result list created by the function run_ica.   \code{ica.result}}

\item{prefix}{Output filename prefix. The output file will be named
"prefix_ICA_summary.html". \code{prefix}}

\item{output.path}{Directory path for generating the output HTML file.
default is set to current working directory. \code{output.path}}
}
\value{
output HTML report that
}
\description{
Generating a HTML report from a ICA list object.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

