% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pc_report}
\alias{pc_report}
\title{Automated PCA report generator}
\usage{
pc_report(phenotype.mx = NULL, info.df = NULL, check.covars = NULL,
  cor.threshold = 0.05, prefix = NULL, output.path = NULL,
  file.ext = "png")
}
\arguments{
\item{info.df}{Sample information dataframe to be used for covariate correlation \code{info.df}}

\item{prefix}{Output filename prefix. The output file will be named
"prefix_ICA_summary.html". \code{prefix}}

\item{output.path}{Directory path for generating the output HTML file.
default is set to current working directory. \code{output.path}}

\item{pca.result}{PCA result list created by the function prcomp().   \code{pca.result}}
}
\value{
output HTML report that contains component wise visualization and summary plots.
}
\description{
Generating a HTML report from a PCA object created with the function prcomp().
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

