% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_coeff_w_legend}
\alias{plot_coeff_w_legend}
\title{Plotting the coefficients for a single IC}
\usage{
plot_coeff_w_legend(info_input, k.col)
}
\arguments{
\item{info_input}{A dataframe that has the index and values of ICs with covariate information saved.}

\item{k.col}{The column name to use for plot coloring.}
}
\value{
ggplot object
}
\description{
ggplot function for plotting coefficients for a single IC
}
\keyword{keywords}

