% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_component}
\alias{plot_component}
\title{Plotting the gene weights of a single component}
\usage{
plot_component(s, plot.title, peaks = FALSE, peakresult = NULL)
}
\arguments{
\item{s}{A single component}

\item{plot.title}{Character title of the plot}

\item{peaks}{TRUE shows peaks in red, FALSE leaves everything black}

\item{peakresult}{Pre-calculated peak positions, if not supplied peaks will be
defined as values lying out of 2 standard deviations.}
}
\value{
ggplot object
}
\description{
ggplot function for plotting a single component
}
\keyword{keywords}

