% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pre_process_data}
\alias{pre_process_data}
\title{Pre-processing of the data}
\usage{
pre_process_data(phenotype.mx, scale.pheno = TRUE)
}
\arguments{
\item{phenotype.mx}{Phenotype matrix with a dimension of g X N  \code{phenotype.mx}}
}
\value{
Pre-processed phenotype matrix
}
\description{
The function takes in the phenotype matrix
and removes any 0 variance entries. Also it mean centers the data.
If scaling option is true it will scale all the rows to have unit standard deviation.
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

