% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_ica}
\alias{run_ica}
\title{Automated ICA analysis with Report}
\usage{
run_ica(phenotype.mx = NULL, info.df = NULL, check.covars = NULL,
  k.est = "default", scale.pheno = FALSE, n.runs = 5, max.iter = 10,
  n.cores = 1, cor.threshold = 0.05, ...)
}
\arguments{
\item{phenotype.mx}{Phenotype matrix   \code{phenotype.mx}}

\item{info.df}{dataframe that holds covariates \code{info.df}}

\item{check.covars}{}

\item{k.est}{}

\item{scale.pheno}{}

\item{n.runs}{}

\item{max.iter}{}

\item{n.cores}{}

\item{cor.threshold}{}
}
\value{
output HTML report that
}
\description{
Performing ICA + Report Generation
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

